Write script to generate hmm files
Reads germline genes from file, and makes stochhmm model file to
distinguish them. The version in this commit just goes to the end of V,
which is great... except it turns out stochhmm can't do silent
states. This is a problem because the global hmm diagram will be *much*
more complicated without silent states, so I'll have to figure out
whether dealing with that, or implementing silent states in stochhmm, is easier.