Clear out unused algorithms
There was an impressive profusion of different flavors of forward,
backward, viterbi, and posterior. It was a bit damp in there, but more
than that, I need to modify them, and I'm not going to modify all
twenty-odd ones, so the ones I don't need get nuked and can come back
later.

These are the ones I'm using:

 2112  ./stochhmm -seq examples/Dice_short.fa -model examples/Dice.hmm -posterior
 2113  ./stochhmm -seq examples/Dice_short.fa -model examples/Dice.hmm -stochastic viterbi -repetitions 3
 2114  ./stochhmm -seq examples/Dice_short.fa -model examples/Dice.hmm -viterbi