Add ability to chop up query sequence
Parametrized with lengths k_v and k_d.

Also tore out a huge amount of sequences class, since the
variable num_of_sequences was conflicting with sequences::size() and
causing things to fail mysteriously and catastrophically.

So with this command

> ./stochhmm -seq bcell/seq.fa -viterbi -label

we get an actual maximum in the probability at the correct k_v, k_d

         295          14    -141.988
         295          15    -141.408
         295          16    -125.427
         295          17    -120.997
         295          18    -81.9161
         295          19    -85.1865
         295          20    -88.4568

Badass.