Put each gene version in separate hmm file
The probabilities can factorize like this, and if you try to put all the
v versions in one file you get a million-line text file.

The other way around it would be to eliminate the text file silliness
altogether and interface the code that writes the text files straight to
the hmm creation code.