rewrite version counter
so you can integrate over any set of variables. The point being: for
input to hmmwriter, we need to know the prob of each combination, taking
into account only the variables with which each is correlated.

Damn, that sentence makes no sense. Um, I can't think of a better way to
say it. sigh.