"Finish" testing k_v k_d command line interface
that I some day hope to call with numbers from s-w

finish in big bad huge quotes 'cause I just discovered that despite
imprecations to the contrary in the documentation, the stochhmm END transition
prob does *not* have to be 1. Rad! That makes *way* more sense, and in
fact if it had had to be 1 would've probably made it impossible to do
shit right.