finish on-the-fly parameter counting *KINDA*
the 'kinda' refers to forgetting to do the mute freqs. But, I finished the
infrastructure to read and write on the fly within the waterer.

Also, disabled cdr3 length partitioning, because it turns out that in
some (maybe one percent?) of query seqs you can't reliably infer the
cdr3 length. I think I will be able to fix this by just allowing the
clusterer to handler uncertainty in the clustering variable (i.e.
this sequence's cdr3 length is either x or y).