swap data files in data/test/
The old one was the 'head -n10000' of human A's merged data... but since
this was sorted, every sequence started with T's. Which has the
potential for some bias.

So, replace it with a file that takes every hundredth line from
the (~five-million-line) data file.

The new files about five times as many lines, but also now bzip2'ed