many changes to get imgt parser to work better
The main problem is imgt frequently adds new genes, changes the sequence
to which a name corresponds, and changes the name to which a sequence
corresponds. Well, and that the output is in a text format that's a
monumental biznitch to parse.

Great, so I can just download their current genes and rerun my whole
shingamadong on that, right? Well that won't take too horribly long, but
then I'm still screwed because joinsolver's web interface doesn't allow
you to specify the input germlines. Oh, and I can't download their
current versions anyway 'cause their fucking web site is kicking java
exceptions.