fix random seed problem in recombinator
Now that, senor ralph, is a solid bug.

I was giving bppseqgen os.getpid() as a random seed... which was ok in
the past when I was running separate procs for each reco event. But now
I'm not, so the tree for every rearrangement event was getting simulated
with the same seed. In my defense, it turns out that the mutation
frequency distribution you get from running with all the same seeds is
spiky in exactly the same way as the *actual* mute freq distribution
from data. The spikes are even in the same places (although the data
spike distribution is convolved with the spike distribution from the
random seed problem)... it's only clear there's a problem if you pass in
a uniform distribution.