fix end-trans log prob in chunk caching
arg, that was wrong. I wasn't selecting the best viterbi value *after*
multiplying by the log prob of an end transition, but instead was
finding the best without it and then multiplying afterward. Didn't
notice until I was running bcell stuff because the end-trans probs were
the same for both states in the cpg example.