increase default v fuzz
From 2 to 5. The problem was that for some v genes we only see in data
3p erosions of zero or one base, so the hmm absolutely forbids erosions
longer than one. But the smith-waterman doesn't know about this, so it
passes in k sets that require longer erosions than this, and then the
hmm can't find a valid path.

Note that it's arguable that we should extrapolate to longer erosions
than we observe... but I'm kinda dubious.