fully disable joint/pair emission
having two conflicting ways to decare emissions in the yamels was
just... bad. I don't like that this makes it so you can't specify joint
emission, but you could only specify joint emission for k=2 before,
anyway, and it was confusing. One of the best things about how the k-hmm
is implemented is that it runs cleanly on however many sequences you
pass it, deciding what k was on the fly from the number of sequences,
and the pair-hmm joint table in the yamel totally subverts that
model. So it dies. For now.

NOTE the yamels in bcrham/test/data/regression still have pair
emissions, they're just ignored. I suppose I should add a test that
there aren't tags in the input yamels that I don't use, but it doesn't
seem important at the moment.