switching over to histograms
instead of rounding in mutefreqer

DEAR FUTURE SELF those weird spikes in the all-mean-mute-freqs plot that
are not in the the v-mean-mute-freqs plot that you keep freaking out
over are fine. They're from aliasing/binning interactions between the
fixed sequence length in 'all' (130bp) and the binning in the histogram.