update test/data/regression
this process is getting lame. and just diffing doesn't really do
anything do anything to tell me if anything's broken, since we're still
at the stage where we're making loads of changes that we expect to
change things. So in the end you're just looking at plots, which you
can't really automate. sigh.