remove insertion base content option
it was turned on by default any way

I don't know what the comment in partis.py about it slowing down things
by a factor of five was about. Maybe it was applicable to an old hmm
topology, but I don't think as currently set up non-1/4 probs will
be any slower.