switch naive seqs to same name translator as log probs
i.e. stop replacing a query with a parent if the parent's bigger than
20, and instead always calculate it, but with a subset of the sequences.
Should be fine, as long as the subset is reasonably small it'll still be super fast.

Note, this could make point partis much more accurate (i.e. note the several, large,
over-merged clusters in some of the plots)